+++ Changelog.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 Be aware that code written for 1.9 will not work out of the box because DotSpatial.Topology was replaced by DotSpatial.GeoAPI and DotSpatial.NetTopologySuite (#786).
 
 ### Added
+- Aliases to Satellite of DotSpatial.Positioning
 - Switched to NTS/GeoAPI instead of DotSpatial.Topology (#633, #404, #786) 
 - Tag property in ActionItem (#338)
 - Added property AppManager.BaseDirectory which allows to change base directory for plugins. (#758)
@@ -40,6 +41,7 @@ Be aware that code written for 1.9 will not work out of the box because DotSpati
 - Removed DotSpatial.Mono assembly. Mono helper now is in DotSpatial.Data assembly.
 
 ### Fixed
+- Satellite's missing properties (#958)
 - Parameters for the Austrian Bundesmeldenetz in DotSpatial.Projections.ProjectedCategories.NationalGrids are incorrect (#855)
 - Raster extent shifts from correct extent (#725)
 - Inconsistent use of affine coefficients (#822)